+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 2.6.7
+### 2.7.0
 
 - Adds `header.title` and `header.subtitle` YAML config to customize Cortex page title and subtitle
 - Adds `hideCortexLinks` YAML flag to hide all links to Cortex app